#79 ロジック修正
